Restructure ur-game.lisp
The program needs a huge rework to separate between the game engine and
the web app. This commit reorders quite a bit from the web app logic,
and moves logic from draw offering to the game engine. The idea is that
the only state the web app should know about the client is its color.