Remove src/config.lisp
It's not needed for now while I tear down deployment to work on
refactoring. There might be a better way to deploy, or there might not
be an explicit way entirely.

Main concern right now is how it would find where the static docs are.
Let's work on that in the future.