Overhaul path segment algorithm
The original path algorithm broke the two player routes as five separate
arrays, and functions operating on them had to smartly decide based on
the index which one to use.

Now, those five arrays are stored in two different types of data
structures: the game board, which is for data serialization, and two
instances of a special type of linked list, with one "next" slot for
each player.